Move operator_util.py into utils/
*OSBS-8605

Also change files which import operator_util.py
